+++ src/DailySports.Backend/Controllers/GamesController.cs
@@ -33,7 +33,7 @@ namespace DailySports.Backend.Controllers
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public IActionResult Create([Bind("Id,Name,GameImage")] Game game,IFormFile file)
+        public IActionResult Create(Game game,IFormFile file)
         {
             if (ModelState.IsValid)
             {
@@ -70,13 +70,14 @@ namespace DailySports.Backend.Controllers
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public IActionResult Edit([Bind("Id,Name,GameImage")] Game game, IFormFile file, string oldFileName)
+        public IActionResult Edit(Game game, IFormFile file, string oldFileName)
         {
+            game.GameImage = oldFileName;
             if (ModelState.IsValid)
-            {
+            { 
                 if (file != null)
                 {
-                    if (oldFileName.Length > 0)
+                    if (oldFileName != null && oldFileName.Length > 0)
                     {
                         GoogleStorageService.Delete(oldFileName);
                     }
@@ -86,6 +87,7 @@ namespace DailySports.Backend.Controllers
                 db.SaveChanges();
                 return RedirectToAction("Index");
             }
+            ViewBag.oldFileName = game.GameImage;
             return View(game);
         }
