+++ src/DailySports.Backend/Controllers/PetOfTheDaysController.cs
@@ -33,7 +33,7 @@ namespace DailySports.Backend.Controllers
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public IActionResult Create([Bind("Id,Title,Description,PetImage,Age,Gender,FunFact,Owner,StartDate,EndDate")] PetOfTheWeek petOfTheWeek, IFormFile file)
+        public IActionResult Create(PetOfTheWeek petOfTheWeek, IFormFile file)
         {
             if (ModelState.IsValid)
             {
@@ -70,13 +70,14 @@ namespace DailySports.Backend.Controllers
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public IActionResult Edit([Bind("Id,Title,Description,PetImage,Age,Gender,FunFact,Owner,StartDate,EndDate")] PetOfTheWeek petOfTheWeek, IFormFile file, string oldFileName)
+        public IActionResult Edit(PetOfTheWeek petOfTheWeek, IFormFile file, string oldFileName)
         {
+            petOfTheWeek.PetImage = oldFileName;
             if (ModelState.IsValid)
             {
                 if (file != null)
                 {
-                    if (oldFileName.Length > 0)
+                    if (oldFileName != null && oldFileName.Length > 0)
                     {
                         GoogleStorageService.Delete(oldFileName);
                     }
@@ -86,6 +87,7 @@ namespace DailySports.Backend.Controllers
                 db.SaveChanges();
                 return RedirectToAction("Index");
             }
+            ViewBag.oldFileName = petOfTheWeek.PetImage;
             return View(petOfTheWeek);
         }
